server-client test
put together a very basic server-client test, it's very rough and i had
to use a couple work arounds because im terrible at server stuff

basic overview of how it works:
- client and server connect on the same ports
- client sends a message to server
- server asks the protocol what to do
- currently the protocol only supports parsing the input as a att
(action target target), but in the future it'll support login stuff too
probably
- server gets something back from the protocol
- server passes that back to the client
- client prints whatever it got from server

right now the server protocol takes a string[] from parse, converts it
to a string with #s so that it can be send as a string to the client,
and the client splits the string based on #'s and prints it. it's
horribly messy and i need to figure out on how to pass strings better.

also to run:
-> import projects into eclipse or other IDE that supports jar building
-> export jar for the server
-> export jar for the client in apk/main/client/client.java, not the old
client
-> open up cmd in the folder where you exported the jars
-> start server first by going:
java -jar [filenameofServer].jar [port]

eg.
java -jar apkServer.jar 4000

-> then start the client in the same manner but with a host
java -jar [filenameofClient].jar [hostname] [port]

eg.
java -jar apkClient.jar localhost 4000