messing with openGL
messed around with openGL, thought it might be good to work a little on
a graphical side of my currently text-only engine

learned a lot of stuff about openGL:
- coordinates are in the bottom left instead of the top right because
GRAPHICS
-> this means that there's no sane way to render maps with my engine,
since stuff is printed top-to-bottom and not bottom to top

i tried figuring out the coordinates business but either i'm dumb or
opengl is messing with me because i couldn't get the box and mouse
coordinates to match up (unless i rendered everything upsidedown)

maybe its just slick-utils fault for assuming that everyone would always
flip their text and/or render it top-left to bottom-right. maybe its not
upsidedown if i set glOrtho(0,800,0,600,1,-1);, maybe a lot of things

point is i dont find openGL very viable given that i've already made a
pretty terrible text based adventure game thing (where a lot of the
issues i encounter are because my engine is just too convoluted to use)

next stuff to mess with for fun: unity? maybe.